Stock Market Management System
It is a web project in which we pull several units from the stock exchange via API and at the same time establish a database connection with user login.
PHP was used during development.
In the project, each member can create their own account and add to their wallet by buying and selling transactions at current foreign currency values. Transactions are not real.